+++ modules/slo-pipeline/variables.tf
@@ -20,16 +20,7 @@ variable "project_id" {
 
 variable "exporters" {
   description = "SLO export destinations config"
-  type        = list(any)
-
-  # wait on https://github.com/hashicorp/terraform/issues/22449 to be merged
-  # type = list(object({
-  #   class = string
-  #   project_id = string
-  #   dataset_id = string
-  #   table_id = string
-  #   topic_name = string
-  # }))
+  type        = any
 }
 
 variable "pubsub_topic_name" {
@@ -52,6 +43,26 @@ variable "function_memory" {
   default     = 128
 }
 
+variable "function_timeout" {
+  description = "Timeout (in seconds)"
+  default     = "90"
+}
+
+variable "function_environment_variables" {
+  description = "Cloud Function environment variables"
+  default     = {}
+}
+
+variable "vpc_connector" {
+  description = "VPC Connector. The format of this field is projects/*/locations/*/connectors/*."
+  default     = null
+}
+
+variable "vpc_connector_egress_settings" {
+  description = "VPC Connector Egress Settings. Allowed values are ALL_TRAFFIC and PRIVATE_RANGES_ONLY."
+  default     = null
+}
+
 variable "region" {
   description = "Region for the App Engine app"
   default     = "us-east1"
@@ -62,7 +73,7 @@ variable "storage_bucket_location" {
   default     = "US"
 }
 
-variable "storage_bucket_storage_class" {
+variable "storage_bucket_class" {
   description = "The Storage Class of the new bucket. Supported values include: STANDARD, MULTI_REGIONAL, REGIONAL, NEARLINE, COLDLINE"
   default     = "STANDARD"
 }
@@ -108,5 +119,19 @@ variable "dataset_create" {
 
 variable "slo_generator_version" {
   description = "SLO generator library version"
-  default     = "1.1.2"
+  default     = "1.4.0"
+}
+
+variable "extra_files" {
+  description = "Extra files to add to the Google Cloud Function code"
+  default     = []
+  type = list(object({
+    content  = string,
+    filename = string
+  }))
+}
+
+variable "labels" {
+  description = "Labels to apply to all resources created"
+  default     = {}
 }